+++ src/hpp/world_init.hpp
@@ -32,7 +32,8 @@ Entity createCrosshair(Entity origin, bool iscat);
 
 Entity createHealthCounter(Entity origin, int health, TextManager& textManager);
 
-Entity createTimerCounter(int newtimer, TextManager& textManager);
+// change option timer type
+Entity createTimerCounter(float newtimer, TextManager& textManager);
 
 Entity createPlayersCounter(int newplayers, TextManager& textManager);
